refactor: 입출력 관련 코드를 상황에 맞게 분리
- 입력과 출력 형태로 구분하던 방식에서 각 상황에서의 입출력으로 분리합니다.
