+++ src/main/java/blackjack/Application.java
@@ -1,13 +1,17 @@
 package blackjack;
 
 import blackjack.controller.BlackjackController;
-import blackjack.view.InputView;
-import blackjack.view.OutputView;
+import blackjack.view.GamePlayView;
+import blackjack.view.GameResultView;
+import blackjack.view.GameSetupView;
 
 public class Application {
 
     public static void main(String[] args) {
-        BlackjackController blackjackController = new BlackjackController(new InputView(), new OutputView());
-        blackjackController.run();
+        createController().run();
+    }
+
+    private static BlackjackController createController() {
+        return new BlackjackController(new GameSetupView(), new GamePlayView(), new GameResultView());
     }
 }
