+++ src/main/java/blackjack/view/GameSetupView.java
@@ -6,11 +6,25 @@ import java.util.List;
 import java.util.Scanner;
 import java.util.regex.Pattern;
 
-public class InputView {
+public class GameSetupView {
 
+    private static final String GAME_BANNER = """
+            
+            ================================
+             Welcome to the Blackjack Game!
+            ================================
+                게임에 참여할 플레이어 이름은
+               2~10글자의 한글, 숫자만 허용되며,
+               다른 플레이어와 중복될 수 없습니다.
+            ================================
+            """;
     private static final Pattern COMMA_WITH_OPTIONAL_SPACES_PATTERN = Pattern.compile("\\s*,\\s*");
 
     private final Scanner scanner = new Scanner(System.in);
+
+    public void printStartBanner() {
+        System.out.println(GAME_BANNER);
+    }
 
     public List<String> readPlayerNames() {
         System.out.println("게임에 참여할 사람의 이름을 입력하세요.(쉼표 기준으로 분리)");
@@ -33,10 +47,5 @@ public class InputView {
         } catch (NumberFormatException e) {
             throw new IllegalArgumentException("숫자만 입력할 수 있습니다. 입력: %s".formatted(input));
         }
-    }
-
-    public ParticipantAction readHitOrNot(Name playerName) {
-        System.out.printf("%s는 한장의 카드를 더 받겠습니까?(예는 y, 아니오는 n)%n", playerName.value());
-        return ParticipantAction.from(scanner.nextLine().trim());
     }
 }
